refactor(app): create useJoinMeeting custom hook
